-rename project -add binary, ascii encoder and decoder -add cmake
