+++ Cryptosharkist/Cryptosharkist.rc
@@ -53,7 +53,7 @@ END
 
 // Icon with lowest ID value placed first to ensure application icon
 // remains consistent on all systems.
-IDI_ICON1               ICON                    "C:\\Users\\Admin\\Downloads\\shark-_1_.ico"
+IDI_ICON1               ICON                    "E:\\Cryptosharkist\\Cryptosharkist\\Cryptosharkist\\shark.ico"
 
 #endif    // English (United States) resources
 /////////////////////////////////////////////////////////////////////////////
