+++ Cryptosharkist/anyascii.c
