+++ Cryptosharkist/anyascii.h
